Create nav_item.html
This code is for a different type of navigation bar or header item. You can customize it yourself. No copyrights are here by me. The owner of this code is "https://codepen.io/Khaosmuhaha." I'm working on it and will customize it myself for the future. This repository is just for me.